requirements.txt
bs4
pandas
MySQL-python
statsmodels.formula.api
